+++ src/crosscutting/util/file_util.cpp
@@ -6,10 +6,10 @@ namespace emu::util::file {
 
     std::vector<u8> convert_char_pointer_to_vector(char *memory, std::streampos size) {
         std::vector<u8> program;
-        program.reserve(size);
+        program.reserve(static_cast<unsigned long>(size));
 
         for (int i = 0; i < size; ++i) {
-            program.push_back(memory[i]);
+            program.push_back(static_cast<u8>(memory[i]));
         }
 
         return program;
@@ -25,7 +25,7 @@ namespace emu::util::file {
 
         if (file.is_open()) {
             size = file.tellg();
-            memory = new char[size];
+            memory = new char[static_cast<unsigned long>(size)];
             file.seekg(0, std::ios::beg);
             file.read(memory, size);
             file.close();
